+++ next/src/pages/[...path].tsx
@@ -5,10 +5,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import * as React from 'react'
 import { useEffect, useMemo } from 'react'
 
+import Breadcrumbs from '@/src/components/common/Breadcrumbs/Breadcrumbs'
 import PageHeaderSections from '@/src/components/layout/PageHeaderSections'
 import SectionContainer from '@/src/components/layout/Section/SectionContainer'
 import Sections from '@/src/components/layout/Sections'
-import BreadcrumbsPlaceholder from '@/src/components/placeholder/BreadcrumbsPlaceholder'
 import PageLayoutPlaceholder from '@/src/components/placeholder/PageLayoutPlaceholder'
 import { GeneralContextProvider } from '@/src/providers/GeneralContextProvider'
 import { client } from '@/src/services/graphql'
@@ -124,7 +124,7 @@ const Page = ({ page, general }: PageProps) => {
       <PageLayoutPlaceholder>
         {/* TODO consider extracting to PageContent */}
         <SectionContainer background="secondary">
-          <BreadcrumbsPlaceholder breadcrumbs={breadcrumbs} />
+          <Breadcrumbs breadcrumbs={breadcrumbs} />
         </SectionContainer>
         <PageHeaderSections header={header} />
